backend/pyproject.toml
[project]
name = "deer-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-sandbox>=0.0.19",
    "dotenv>=0.9.9",
    "langchain>=1.2.3",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "markdownify>=1.2.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "readabilipy>=0.3.0",
    "tavily-python>=0.7.17",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.14.11",
]
